Generate correct markup for selected options.
Previously, each <option> had a "selected=..." attribute, with its value
set to true or false depending on whether it was supposed to be
selected.  Browsers interpret this to mean that we want them all to be
selected.  The correct behavior is to leave out the "selected" attribute
for all but the element that is to be selected.