Allow setting arbitrary properties on the html5 root element
With this change, the options arguemnt of the html5 macro can contain
any option, not just :lang, and they all will be used to form the
properties of the <html> element (except for the :xml? option).

This is accomplished by using defelem for xhtml-tag, and later in the
html5 macro, passing options (sans :xml?) to it as-is.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>